Allow crate to compile without leaking features from dev-dependencies.
When adding blanket as a dependency to a blank cargo project, it fails to
compile. This is due to missing features on the syn crate.

This commit adds the needed features. This problem could not be detected
from within the crate itself because features were leaking through from
the dev-dependency of syn which does not have `default-features = false`.

I turned on `resolver = "2"`. This is the new feature resolver for cargo
that will become the default in edition 2021. This is not strictly necessary
but it helps as it no longer leaks features. This means that `cargo check`
will now fail if there are missing features. As `cargo test` obviously turns
on the features from `dev-dependencies`, this will still not detect this
issue.

You might want to check your CI script and add a `cargo check` before
running `cargo test`.